+++ release/githubio/js/master_uncompressed.js
@@ -7526,7 +7526,7 @@
 
         var changelogObj = [{
                 title: "Web Version September 01 2020",
-                changes: ["Improve Stay on the stream feature, prevent buffer dialog and last video frame from be displayed when the mode starts"]
+                changes: ["Improve Stay on the stream feature, prevent buffer dialog and last video frame from be displayed when the mode starts, prevent screen save"]
             },
             {
                 title: "Web Version August 30 2020",
@@ -7554,13 +7554,6 @@
             {
                 title: "Apk Version 3.0.244 - Web Version August 17 2020",
                 changes: ["General performance improves and bug fixes"]
-            },
-            {
-                title: "Apk Version 3.0.241 - Web Version August 15 2020",
-                changes: [
-                    "Mitigate the freeze effects when using accessibility service, most noticeably when exiting the player, slowdowns do to accessibility service can't be fixed only mitigated, read more about it https://tinyurl.com/applag",
-                    "General performance improves and bug fixes"
-                ]
             },
         ];
 
@@ -10587,6 +10580,13 @@
         } catch (e) {}
     }
 
+    //public void mKeepScreenOn(boolean keepOn)
+    //Android specific: true
+    //Allows to control if the screen will be on or not from js side
+    function OSInterface_mKeepScreenOn(keepOn) { //Not be used
+        Android.mKeepScreenOn(keepOn);
+    }
+
     //public boolean isKeyboardConnected()
     //Android specific: true
     //informs if a hw Keyboard is connected to the devices
@@ -10606,13 +10606,6 @@
     //return the playback state
     // function OSInterface_getPlaybackState() {//Not be used
     //     return Android.getPlaybackState();
-    // }
-
-    //public void mKeepScreenOn(boolean keepOn)
-    //Android specific: true
-    //Allows to control if the screen will be on or not from js side
-    // function OSInterface_mKeepScreenOn(keepOn) {//Not be used
-    //     Android.mKeepScreenOn(keepOn);
     // }
 
     //public void mSetAudio(int position, float volume)
@@ -12153,6 +12146,7 @@
         if (!ChatLive_loaded[0]) ChatLive_Init(0);
         Play_HideBufferDialog();
         if (Main_IsOn_OSInterface) OSInterface_stopVideo();
+        OSInterface_mKeepScreenOn(true);
         Play_showChat();
         Play_data.watching_time = new Date().getTime();
         Play_state = Play_STATE_PLAYING;